+++ packages/core/src/lib/Modal/Modal.tsx
@@ -19,6 +19,7 @@ const Modal: React.FC<ModalProps> = ({
     children,
     closeButton,
     customFooter,
+    customize,
     fixedMode = false,
     hasCancel = true,
     hasFooter = true,
@@ -60,6 +61,7 @@ const Modal: React.FC<ModalProps> = ({
 
     return (
         <DivStyled
+            customize={customize}
             data-testid="test-modal"
             id={id}
             isCentered={isCentered}
@@ -67,12 +69,17 @@ const Modal: React.FC<ModalProps> = ({
             zIndex={zIndex}
             {...props}
         >
-            <DivStyledWrap width={width} canOverflow={canOverflow}>
+            <DivStyledWrap
+                canOverflow={canOverflow}
+                customize={customize}
+                width={width}
+            >
                 <HeaderStyled
+                    customize={customize}
                     data-testid="test-modal-header"
-                    title={title}
                     fixedMode={fixedMode}
                     headerHasBottomBorder={headerHasBottomBorder}
+                    title={title}
                 >
                     <>
                         {typeof title == 'string' ? (
@@ -112,7 +119,11 @@ const Modal: React.FC<ModalProps> = ({
                     </>
                 </HeaderStyled>
 
-                <DivStyledContent id="content" data-testid="test-modal-content">
+                <DivStyledContent
+                    customize={customize}
+                    data-testid="test-modal-content"
+                    id="content"
+                >
                     {children}
                 </DivStyledContent>