+++ packages/icons/src/lib/icons/CheckCircleAlt.tsx
@@ -1,0 +1,42 @@
+// Auto-generated file created by svgr-cli source svg-template.js
+// Run pnpm icons:create to update
+// Do not edit
+import { SVGProps } from 'react';
+import { SVGRProps } from '../type';
+const SvgCheckCircleAlt = ({
+    title,
+    titleId,
+    isResponsive = false,
+    style,
+    ...props
+}: SVGProps<SVGSVGElement> & SVGRProps) => (
+    <svg
+        width="1em"
+        height="1em"
+        viewBox="0 0 24 24"
+        fill="currentColor"
+        xmlns="http://www.w3.org/2000/svg"
+        role="img"
+        data-testid="test-icon"
+        aria-hidden="true"
+        style={
+            isResponsive
+                ? style
+                : {
+                      flex: 'none',
+                      ...style,
+                  }
+        }
+        aria-labelledby={titleId}
+        {...props}
+    >
+        {title ? <title id={titleId}>{title}</title> : null}
+        <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M15.663 3.773A9 9 0 1 0 21 12v-.919a1 1 0 1 1 2 0V12A11.002 11.002 0 0 1 8.188 22.313a11 11 0 1 1 8.289-20.366 1 1 0 1 1-.814 1.826Zm7.044-.48a1 1 0 0 1 0 1.414l-10 10.01a1 1 0 0 1-1.414 0l-3-3a1 1 0 1 1 1.414-1.414L12 12.595l9.293-9.302a1 1 0 0 1 1.414 0Z"
+        />
+    </svg>
+);
+
+export default SvgCheckCircleAlt;